refactor ItemListenerTest
